travel-butler/services/api/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
pydantic>=2.10.0,<3.0
pydantic-settings>=2.6.0,<3.0
supabase>=2.11.0,<3.0
python-jose[cryptography]>=3.3.0,<4.0
httpx>=0.28.0,<1.0
python-multipart>=0.0.18
ruff>=0.8.0
